Docker is OK, relative path
Re-wrote the module to make it so that we do not rely
on the git binary.

I ran into a case where calling the git binary would give me the
following message:

Git clone: error: chmod on <ntfs mount> failed: Operation not permitted

Since git config is written in ini format, simply using
an ini file parser solved the issue.

This commit also solve an issue regarding the use
of relative paths.